+++ src/cljs/c3kit/wire/google.cljs
@@ -1,8 +1,7 @@
 (ns c3kit.wire.google
   (:require [c3kit.apron.corec :as ccc]
             [c3kit.apron.log :as log]
-            [c3kit.wire.js :as wjs]
-            [reagent.core :as reagent]))
+            [c3kit.wire.js :as wjs]))
 
 ;; https://developers.google.com/identity/gsi/web/reference/html-reference#element_with_class_g_id_signin
 
@@ -14,14 +13,14 @@
 (defn- on-button-mount [options node]
   (if (wjs/doc-ready?)
     (mount-oauth-button node options)
-    (wjs/add-listener js/window "load" #(mount-oauth-button node options))))
+    (wjs/add-listener js/window "load" #(mount-oauth-button node options) :once true)))
 
-(defn- with-ref [[tag maybe-options :as body] ref]
+(defn- with-ref [[tag maybe-options :as body] ref-fn]
   (let [has-options? (map? maybe-options)
-        ref-option   (:ref maybe-options)
-        ref-option   (cond-> #(reset! ref %)
-                             (and has-options? ref-option)
-                             (juxt ref-option))
+        user-ref     (:ref maybe-options)
+        ref-option   (if (and has-options? user-ref)
+                       (fn [node] (ref-fn node) (user-ref node))
+                       ref-fn)
         options      (cond->> {:ref ref-option}
                               has-options?
                               (merge maybe-options))
@@ -32,8 +31,7 @@
 
 (defn oauth-button
   "Renders the Google OAuth Button."
-  [options _body]
-  (let [node (atom nil)]
-    (reagent/create-class
-      {:component-did-mount #(on-button-mount options @node)
-       :reagent-render      (fn [_options body] (with-ref body node))})))
+  [options body]
+  (with-ref body (fn [node]
+                   (when node
+                     (on-button-mount options node)))))